accept patch from an anonymous edging that says mesh code is now complete, but needs improved.  scene nodes can now report if they are would have been freed (and are not freed automatically upon request), treads behave on the physics drivers better, and a lil cleanup.
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@8787 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/geometry/TankSceneNode.cxx
@@ -240,8 +240,8 @@ void TankSceneNode::addRenderNodes(SceneRenderer& renderer)
     GLfloat dx = eye[0] - sphere[0];
     GLfloat dy = eye[1] - sphere[1];
     const float radians = azimuth * M_PI / 180.0f;
-    const float cos_val = cos(radians);
-    const float sin_val = sin(radians);
+    const float cos_val = cosf(radians);
+    const float sin_val = sinf(radians);
 
     const float frontDot = (cos_val * dx) + (sin_val * dy);
     const bool towards = (frontDot > 0.0f);
